Add LocalAuthentication coverage, oracle queue, and README
Record 233 declared / 14 deferred precise IDs from the public surface, keep focused checks in LocalAuthenticationRuntime.swift, and document fail-closed Linux behavior. Combine, Security, and the unobserved reuse-duration constant stay deferred.
