+++ full/localauthentication/oracle-questions.tsv
@@ -1,0 +1,11 @@
+precise	question	risk	reason
+c:@LAErrorDomain	What is the exact NSError domain string Apple's iOS 26.1 LocalAuthentication binary stores in LAErrorDomain?	fail-closed	The seed records the exported symbol and macios ErrorDomain("LAErrorDomain") annotation, not the runtime payload. Linux uses the C identifier as process-local identity.
+c:@LATouchIDAuthenticationMaximumAllowableReuseDuration	What TimeInterval does Apple export for LATouchIDAuthenticationMaximumAllowableReuseDuration on iOS 26.1?	fail-closed	The graph and TBD list the symbol without a numeric value. The property default 0 is documented; the maximum constant is not.
+c:objc(cs)LAContext(im)evaluatePolicy:localizedReason:reply:	On iOS 26.1, does evaluatePolicy invoke its reply inline or after return, and on which queue?	fail-closed	Linux delivers the fail-closed reply synchronously on the calling thread so in-tree guests that count replies immediately keep working. Apple's queue and after-return behavior are unobserved.
+c:objc(cs)LAContext(im)canEvaluatePolicy:error:	Does Apple write an NSError into an NSErrorPointer, and can a Swift LAError? out-pointer be used at the overlay boundary?	fail-closed	Isolated Linux Foundation has no NSErrorPointer typealias. This port keeps UnsafeMutablePointer<LAError?>? from the original lane.
+c:objc(cs)LAContext(im)evaluateAccessControl:operation:localizedReason:reply:	Which LAError.Code does Apple return when evaluateAccessControl is asked to use a SecAccessControl object on a host with no Secure Enclave?	fail-closed	The method cannot be declared without Security.SecAccessControl. Linux omits it rather than shipping a lookalike.
+s:7Combine16ObservableObjectPA2A0bC9PublisherC0c10WillChangeD0RtzrlE06objecteF0AEvp::SYNTHESIZED::c:objc(cs)LAContext	Does Apple's LAContext emit objectWillChange for localizedReason and related stored properties, and on which scheduler?	ui	Combine is not a declared dependency of this seed. Isolated host compilation does not import Combine.
+c:objc(cs)LARightStore(im)saveRight:identifier:completion:	Which error domain and code does Apple use when LARightStore cannot persist a right, and does a later right(forIdentifier:) see that save?	fail-closed	Linux always fails closed with LAError.biometryNotAvailable and never writes a store. Apple Keychain/right-store errors are unobserved.
+c:objc(cs)LAEnvironment(cpy)currentUser	On a device with no enrolled biometry, is LAEnvironment.State.biometry nil or a non-usable MechanismBiometry, and is userPassword present?	fail-closed	Linux reports a non-usable biometry mechanism and an unset user-password mechanism so those types are reachable. Apple's empty-state shape is unobserved.
+s:SC7LAErrorLeV11errorDomainSSvpZ	Does LAError.errorDomain equal LAErrorDomain byte-for-byte on iOS 26.1, including any reverse-DNS payload?	fail-closed	The overlay ties errorDomain to the process-local LAErrorDomain constant. Apple's bridged CustomNSError domain string is unconfirmed.
+module	Must every evaluate/authorize/store path remain fail-closed until an Apple local-authentication service exists on Linux, or may a documented subset succeed with applicationPassword?	fail-closed	Application-password storage on LAContext is in-memory only. Policy evaluation still fails closed so a fabricated device-owner success cannot be claimed.
